requirements.txt
amqp==5.1.1
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
future==0.18.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
kombu==5.2.4
mssql-django==1.0
numpy==1.23.0
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
psutil==5.9.1
PyJWT==1.7.1
pymssql==2.2.5
PyMySQL==1.0.2
pyodbc==4.0.32
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
redis==4.3.3
requests==2.27.1
rfc3986==1.5.0
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
